stm32: unify can_reset() for all controllers.
