+++ lib/stm32/f3/Makefile
@@ -36,7 +36,7 @@ TGT_CFLAGS      += $(DEBUG_FLAGS)
 
 ARFLAGS		= rcs
 
-OBJS		= can.o rcc.o adc.o i2c.o usart.o dma.o flash.o
+OBJS		= rcc.o adc.o i2c.o usart.o dma.o flash.o
 
 OBJS            += gpio_common_all.o gpio_common_f0234.o \
 		   dac_common_all.o usart_common_all.o crc_common_all.o\
